[AXON-1024] Handle Rovo Dev user entitlement check failed
