+++ package.json
@@ -45,7 +45,7 @@
     ],
     "main": "./build/extension/extension",
     "rovoDev": {
-        "version": "0.11.24"
+        "version": "0.11.25"
     },
     "scripts": {
         "vscode:uninstall": "node ./build/extension/uninstall.js",
